+++ cnccode/controller_classes/CTCurrentActivityReport.inc.php
@@ -1165,6 +1165,7 @@ class CTCurrentActivityReport extends CTCNC
                 1
             );
             $totalActivityDurationHours = $dsResults->getValue('totalActivityDurationHours');
+
             $this->template->set_var(
 
                 array(
@@ -1196,10 +1197,10 @@ class CTCurrentActivityReport extends CTCNC
                     ),
                     'engineerName'               => $dsResults->getValue('engineerName'),
                     'customerName'               => $dsResults->getValue('customerName'),
-                    'customerNameDisplayClass'
-                                                 => $this->getCustomerNameDisplayClass(
+                    'customerNameDisplayClass'   => $this->getCustomerNameDisplayClass(
                         $dsResults->getValue('specialAttentionFlag'),
-                        $dsResults->getValue('specialAttentionEndDate')
+                        $dsResults->getValue('specialAttentionEndDate'),
+                        $dsResults->getValue(DBEJProblem::specialAttentionContactFlag)
                     ),
                     'urlViewActivity'            => $urlViewActivity,
                     'linkAllocateAdditionalTime' => $linkAllocateAdditionalTime,
@@ -1391,7 +1392,8 @@ class CTCurrentActivityReport extends CTCNC
     }
 
     function getCustomerNameDisplayClass($specialAttentionFlag,
-                                         $specialAttentionEndDate
+                                         $specialAttentionEndDate,
+                                         $specialAttentionContactFlag
     )
     {
         if (
@@ -1400,6 +1402,11 @@ class CTCurrentActivityReport extends CTCNC
         ) {
             return 'class="specialAttentionCustomer"';
         }
+
+        if ($specialAttentionContactFlag == 'Y') {
+            return 'class="specialAttentionContact"';
+        }
+
         return null;
     }